SAvance con la tabla de usuario

+++ src/main/java/com/Velox/demo/repository/IproductoRepository.java
@@ -3,5 +3,5 @@ package com.Velox.demo.repository;
 import com.Velox.demo.model.producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IproductoRepository extends JpaRepository<producto, Integer> {
+public interface IproductoRepository extends JpaRepository<producto, Long> {
 }
